lang/programming/python/studioclassroom_FulltextSearch/requirements.txt
Flask
Flask-Cors
chardet
mecab-python3
unidic-lite
xmltodict
psycopg2-binary==2.8.6
pymysql==0.9.3
xmltodict
zhconv
jaconv
joblib==1.0.1